+++ subpackages/lib/source/odood/lib/odoo/lodoo.d
@@ -261,6 +261,7 @@ const struct LOdoo {
           **/
         auto runPyScript(in string dbname, in Path script_path) {
             import std.datetime.stopwatch;
+            import std.process: wait;
             enforce!OdoodException(
                 script_path.exists,
                 "Python script %s does not exists!".format(script_path));
@@ -268,16 +269,20 @@ const struct LOdoo {
                 "Running python script %s for databse %s ...",
                 script_path, dbname);
             auto sw = StopWatch(AutoStart.yes);
-            // TODO: may be find some way to display script output when script is still running
-            auto result = run("run-py-script", dbname, script_path.toString);
+
+            // NOTE: This way, script will print on standard stderr/stdout
+            auto pid = runner.addArgs("run-py-script", dbname, script_path.toString)
+                .spawn;
+            // TODO: Find a way to capture script output. Do we need output?
+            auto result = pid.wait;
             sw.stop();
             enforce!OdoodException(
-                result.status == 0,
-                "Python script %s for database %s failed with error (exit-code)!\nOutput: %s".format(
-                    script_path, dbname, result.output));
+                result == 0,
+                "Python script %s for database %s failed with error (exit-code): %s!".format(
+                    script_path, dbname));
             infof(
-                "Python script %s for database %s completed in %s:\nOutput: %s".format(
-                    script_path, dbname, sw.peek, result.output));
+                "Python script %s for database %s completed in %s.".format(
+                    script_path, dbname, sw.peek));
             return result;
         }